Move ProtoCodec to its own subpackage
